# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/users/create": {
            "post": {
                "description": "creates a new user in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new user",
                "operationId": "createUser",
                "parameters": [
                    {
                        "description": "User Creation Data",
                        "name": "userCreationDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserCreationDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created user",
                        "schema": {
                            "$ref": "#/definitions/models.UserResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ErrorDTO": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "The error code",
                    "type": "integer"
                },
                "message": {
                    "description": "The error message",
                    "type": "string"
                }
            }
        },
        "models.UserCreationDTO": {
            "type": "object",
            "properties": {
                "confirm_password": {
                    "description": "Confirmation of the user's password.\n\t@Description:\tThe confirmation of the user's password to ensure accuracy.\n\t@Required:\t\ttrue",
                    "type": "string"
                },
                "email": {
                    "description": "Email address of the user.\n\t@Description:\tThe email address used to identify the user.\n\t@Required:\t\ttrue",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the user.\n\t@Description:\tThe full name of the user.\n\t@Required:\t\ttrue",
                    "type": "string"
                },
                "password": {
                    "description": "Password of the user.\n\t@Description:\tThe password chosen by the user.\n\t@Required:\t\ttrue",
                    "type": "string"
                }
            }
        },
        "models.UserResponseDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}